test_requirements.txt
coveralls==1.0
pytest==2.8.1
pytest-django==2.8.0
flake8
python-social-auth
